+++ src/features/user-level/components/levels.test.tsx
@@ -14,9 +14,9 @@ describe("Levels", () => {
     it("ユーザーレベル情報が取得される", async () => {
       await Levels({ userId: "test-user-id" });
 
-      expect(require("@/lib/services/users").getProfile).toHaveBeenCalledWith(
-        "test-user-id",
-      );
+      expect(
+        require("@/features/user-profile/services/profile").getProfile,
+      ).toHaveBeenCalledWith("test-user-id");
       expect(
         require("@/features/user-level/services/level").getUserLevel,
       ).toHaveBeenCalledWith("test-user-id", undefined);
